Start Scripts & Package Updated

+++ start-test-server.sh
@@ -1,19 +1,24 @@
 #!/bin/bash
 
+if [[ -z "${TEST_SERVER_HOME}" ]]; then
+  echo "Env Variable \"TEST_SERVER_HOME\" not set."
+  exit 1
+fi
+
 # Use a glob to match filenames directly
 oldJarFileNames=()
 while IFS= read -r -d '' file; do
     oldJarFileNames+=("$(basename "${file}")")
 done < <(find "${TEST_SERVER_HOME}/plugins" -maxdepth 1 -type f -name "CDKMinecraft-*.jar" -print0)
 
-echo "Deleting the following old plugin files from ${TEST_SERVER_HOME}/plugins"
+echo "Deleting the following old plugin files from ${TEST_SERVER_HOME}/plugins:"
 i=1
 for file in "${oldJarFileNames[@]}"; do
     echo "[${i}]: ${file}"
     rm -f "${TEST_SERVER_HOME}/plugins/${file}"
     ((i++))  # Increment i
 done
-
+echo
 # Find the latest created file matching the pattern
 jarFileName=$(find ./build/libs -maxdepth 1 -type f -name "CDKMinecraft-*-all.jar" -printf '%T@ %p\n' | sort -n | tail -1 | cut -d' ' -f2)
 
@@ -25,6 +30,6 @@ else
     echo "No matching file found."
     exit 1
 fi
-
+echo
 cd "${TEST_SERVER_HOME}" || exit 1
 ./start.sh
